Code scanning bug fixed

+++ linkis-commons/linkis-common/src/main/java/org/apache/linkis/common/utils/ByteTimeUtils.java
@@ -350,7 +350,7 @@ public class ByteTimeUtils {
       }
     }
 
-    public double toBytes(long d) {
+    public long toBytes(long d) {
       if (d < 0) {
         throw new IllegalArgumentException("Negative size value. Size must be positive: " + d);
       }
